--- a/lib/ex_json_logger.ex
+++ b/lib/ex_json_logger.ex
@@ -25,8 +25,7 @@ defmodule ExJsonLogger do
   @doc """
   Function referenced in the `:format` config.
   """
-  @spec format(Logger.level(), Logger.message(),
-               Logger.Formatter.time(), Keyword.t()) :: String.t()
+  @spec format(Logger.level(), Logger.message(), Logger.Formatter.time(), Keyword.t()) :: iodata()
   def format(level, msg, timestamp, metadata) do
     logger_info = %{
       level: level,
@@ -43,21 +42,30 @@ defmodule ExJsonLogger do
       encode(%{
         level: :error,
         time: format_timestamp(timestamp),
-        msg: "Could not format: #{inspect({level, msg, metadata})}"
+        msg: "#{__MODULE__} could not format: #{inspect({level, msg, metadata})}"
       })
   end
 
   defp encode(log_event) do
-    log_event
-    |> Jason.encode!()
-    |> Kernel.<>("\n")
+    [Jason.encode_to_iodata!(log_event), ?\n]
   end
 
   defp format_timestamp({date, time}) do
-    "#{format_date(date)} #{format_time(time)}"
+    unsafe_fragment([format_date(date), ?\s, format_time(time)])
   end
 
-  defp format_metadata(pid) when is_pid(pid), do: inspect(pid)
-  defp format_metadata(ref) when is_reference(ref), do: inspect(ref)
+  defp format_metadata(pid) when is_pid(pid) do
+    unsafe_fragment(["#PID"  | :erlang.pid_to_list(pid)])
+  end
+
+  defp format_metadata(ref) when is_reference(ref) do
+    ~c"#Ref" ++ rest = :erlang.ref_to_list(ref)
+    unsafe_fragment(["#Reference" | rest])
+  end
+
   defp format_metadata(other), do: other
+
+  defp unsafe_fragment(data) do
+    Jason.Fragment.new([?", data, ?"])
+  end
 end