👋

This PR makes the formatter use iodata where possible which improves performance somewhat:

<details><summary>Before</summary>

```console
$ MIX_ENV=bench mix run bench/format.exs
Compiling 1 file (.ex)
Operating System: macOS
CPU Information: Apple M1
Number of Available Cores: 8
Available memory: 8 GB
Elixir 1.15.7
Erlang 26.1.2

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: debug, info
Estimated total run time: 14 s

Benchmarking format/4 with input debug ...
Benchmarking format/4 with input info ...

##### With input debug #####
Name               ips        average  deviation         median         99th %
format/4      370.19 K        2.70 μs   ±804.32%        2.21 μs        4.50 μs

##### With input info #####
Name               ips        average  deviation         median         99th %
format/4      236.11 K        4.24 μs   ±417.55%        3.71 μs        7.38 μs
```

</details>

<details><summary>After</summary>

```console
$ MIX_ENV=bench mix run bench/format.exs
Compiling 1 file (.ex)
Operating System: macOS
CPU Information: Apple M1
Number of Available Cores: 8
Available memory: 8 GB
Elixir 1.15.7
Erlang 26.1.2

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: debug, info
Estimated total run time: 14 s

Benchmarking format/4 with input debug ...
Benchmarking format/4 with input info ...

##### With input debug #####
Name               ips        average  deviation         median         99th %
format/4      956.69 K        1.05 μs  ±2928.36%        0.92 μs        1.25 μs

##### With input info #####
Name               ips        average  deviation         median         99th %
format/4      616.96 K        1.62 μs  ±1391.15%        1.46 μs        1.88 μs
```

</details>

If this is something you are interested in, I will cleanup the PR and resubmit.