@Stratus3D 2023-12-13
Thanks for the PR @ruslandoga ! Sorry for the late reply here. I wasn't able to get around to investigating this until today.

I see similar numbers on my system as well. It's about twice as fast as before. That's enough that I think we should include these changes in the next version.