@Stratus3D 2023-12-14
Thanks for the PR @ruslandoga ! I'm closing this in favor of #26 , which includes all your changes here.